requirements.txt
aioredis==1.3.0
aiohttp==3.6.2
aiojobs==0.2.2
aiosqlite==0.10.0
aiopg==1.0.0
amqp==2.5.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.3.0
Django==2.2.6
django-jet==1.0.8
djangorestframework==3.10.3
drf-yasg==1.17.0
django-environ==0.4.5
gunicorn==20.0.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
inflection==0.3.1
ipython==7.9.0
itypes==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
kombu==4.6.5
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
packaging==19.2
pycparser==2.19
PyHamcrest==1.9.0
pyparsing==2.4.5
pyrsistent==0.15.5
pyrogram==0.16.0
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
psycopg2-binary==2.8.4
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
tgcrypto==1.2.0
Twisted==19.10.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
virtualenv==16.7.7
zipp==0.6.0
zope.interface==4.7.1
